static server side rendering added
